requirements.txt
cryptography>=41.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
google-generativeai>=0.8.0
yfinance>=0.2.36
pandas>=2.0.0
